move the tests around
