+++ tests/base/test_tensor.py
