+++ tests/base/test_tensor_boundary.py
