+++ tests/base/test_tensor_innerproduct.py
